requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.1.1
django-appconf==1.0.4
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.3.0
django-fs-trumbowyg==0.1.4
django-tinymce-lite==0.0.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.2.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
